# much faster blends, couple other little things