--- a/src/pixie/blends.nim
+++ b/src/pixie/blends.nim
@@ -124,18 +124,18 @@ proc `-`*(c: Color, v: float32): Color {.inline.} =
   result.b = c.b - v
   result.a = c.a - v
 
-proc screen(backdrop, source: float32): float32 {.inline.} =
-  1 - (1 - backdrop) * (1 - source)
+# proc screen(backdrop, source: float32): float32 {.inline.} =
+#   1 - (1 - backdrop) * (1 - source)
 
-proc hardLight(backdrop, source: float32): float32 {.inline.} =
-  if source <= 0.5:
-    backdrop * 2 * source
-  else:
-    screen(backdrop, 2 * source - 1)
+# proc hardLight(backdrop, source: float32): float32 {.inline.} =
+#   if source <= 0.5:
+#     backdrop * 2 * source
+#   else:
+#     screen(backdrop, 2 * source - 1)
 
-proc softLight(backdrop, source: float32): float32 {.inline.} =
-  ## Pegtop
-  (1 - 2 * source) * backdrop ^ 2 + 2 * source * backdrop
+# proc softLight(backdrop, source: float32): float32 {.inline.} =
+#   ## Pegtop
+#   (1 - 2 * source) * backdrop ^ 2 + 2 * source * backdrop
 
 proc Lum(C: Color): float32 {.inline.} =
   0.3 * C.r + 0.59 * C.g + 0.11 * C.b
@@ -183,103 +183,103 @@ proc alphaFix(backdrop, source, mixed: Color): Color =
   result.g /= result.a
   result.b /= result.a
 
-proc blendNormalFloats*(backdrop, source: Color): Color {.inline.} =
-  result = source
-  result = alphaFix(backdrop, source, result)
+# proc blendNormalFloats*(backdrop, source: Color): Color {.inline.} =
+#   result = source
+#   result = alphaFix(backdrop, source, result)
 
-proc blendDarkenFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = min(backdrop.r, source.r)
-  result.g = min(backdrop.g, source.g)
-  result.b = min(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendDarkenFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = min(backdrop.r, source.r)
+#   result.g = min(backdrop.g, source.g)
+#   result.b = min(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendMultiplyFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = backdrop.r * source.r
-  result.g = backdrop.g * source.g
-  result.b = backdrop.b * source.b
-  result = alphaFix(backdrop, source, result)
+# proc blendMultiplyFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = backdrop.r * source.r
+#   result.g = backdrop.g * source.g
+#   result.b = backdrop.b * source.b
+#   result = alphaFix(backdrop, source, result)
 
-proc blendLinearBurnFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = backdrop.r + source.r - 1
-  result.g = backdrop.g + source.g - 1
-  result.b = backdrop.b + source.b - 1
-  result = alphaFix(backdrop, source, result)
+# proc blendLinearBurnFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = backdrop.r + source.r - 1
+#   result.g = backdrop.g + source.g - 1
+#   result.b = backdrop.b + source.b - 1
+#   result = alphaFix(backdrop, source, result)
 
-proc blendColorBurnFloats*(backdrop, source: Color): Color {.inline.} =
-  proc blend(backdrop, source: float32): float32 {.inline.} =
-    if backdrop == 1:
-      1.0
-    elif source == 0:
-      0.0
-    else:
-      1.0 - min(1, (1 - backdrop) / source)
-  result.r = blend(backdrop.r, source.r)
-  result.g = blend(backdrop.g, source.g)
-  result.b = blend(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendColorBurnFloats*(backdrop, source: Color): Color {.inline.} =
+#   proc blend(backdrop, source: float32): float32 {.inline.} =
+#     if backdrop == 1:
+#       1.0
+#     elif source == 0:
+#       0.0
+#     else:
+#       1.0 - min(1, (1 - backdrop) / source)
+#   result.r = blend(backdrop.r, source.r)
+#   result.g = blend(backdrop.g, source.g)
+#   result.b = blend(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendLightenFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = max(backdrop.r, source.r)
-  result.g = max(backdrop.g, source.g)
-  result.b = max(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendLightenFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = max(backdrop.r, source.r)
+#   result.g = max(backdrop.g, source.g)
+#   result.b = max(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendScreenFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = screen(backdrop.r, source.r)
-  result.g = screen(backdrop.g, source.g)
-  result.b = screen(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendScreenFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = screen(backdrop.r, source.r)
+#   result.g = screen(backdrop.g, source.g)
+#   result.b = screen(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendLinearDodgeFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = backdrop.r + source.r
-  result.g = backdrop.g + source.g
-  result.b = backdrop.b + source.b
-  result = alphaFix(backdrop, source, result)
+# proc blendLinearDodgeFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = backdrop.r + source.r
+#   result.g = backdrop.g + source.g
+#   result.b = backdrop.b + source.b
+#   result = alphaFix(backdrop, source, result)
 
-proc blendColorDodgeFloats*(backdrop, source: Color): Color {.inline.} =
-  proc blend(backdrop, source: float32): float32 {.inline.} =
-    if backdrop == 0:
-      0.0
-    elif source == 1:
-      1.0
-    else:
-      min(1, backdrop / (1 - source))
-  result.r = blend(backdrop.r, source.r)
-  result.g = blend(backdrop.g, source.g)
-  result.b = blend(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendColorDodgeFloats*(backdrop, source: Color): Color {.inline.} =
+#   proc blend(backdrop, source: float32): float32 {.inline.} =
+#     if backdrop == 0:
+#       0.0
+#     elif source == 1:
+#       1.0
+#     else:
+#       min(1, backdrop / (1 - source))
+#   result.r = blend(backdrop.r, source.r)
+#   result.g = blend(backdrop.g, source.g)
+#   result.b = blend(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendOverlayFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = hardLight(source.r, backdrop.r)
-  result.g = hardLight(source.g, backdrop.g)
-  result.b = hardLight(source.b, backdrop.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendOverlayFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = hardLight(source.r, backdrop.r)
+#   result.g = hardLight(source.g, backdrop.g)
+#   result.b = hardLight(source.b, backdrop.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendHardLightFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = hardLight(backdrop.r, source.r)
-  result.g = hardLight(backdrop.g, source.g)
-  result.b = hardLight(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendHardLightFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = hardLight(backdrop.r, source.r)
+#   result.g = hardLight(backdrop.g, source.g)
+#   result.b = hardLight(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendSoftLightFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = softLight(backdrop.r, source.r)
-  result.g = softLight(backdrop.g, source.g)
-  result.b = softLight(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendSoftLightFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = softLight(backdrop.r, source.r)
+#   result.g = softLight(backdrop.g, source.g)
+#   result.b = softLight(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendDifferenceFloats*(backdrop, source: Color): Color {.inline.} =
-  result.r = abs(backdrop.r - source.r)
-  result.g = abs(backdrop.g - source.g)
-  result.b = abs(backdrop.b - source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendDifferenceFloats*(backdrop, source: Color): Color {.inline.} =
+#   result.r = abs(backdrop.r - source.r)
+#   result.g = abs(backdrop.g - source.g)
+#   result.b = abs(backdrop.b - source.b)
+#   result = alphaFix(backdrop, source, result)
 
-proc blendExclusionFloats*(backdrop, source: Color): Color {.inline.} =
-  proc blend(backdrop, source: float32): float32 {.inline.} =
-    backdrop + source - 2 * backdrop * source
-  result.r = blend(backdrop.r, source.r)
-  result.g = blend(backdrop.g, source.g)
-  result.b = blend(backdrop.b, source.b)
-  result = alphaFix(backdrop, source, result)
+# proc blendExclusionFloats*(backdrop, source: Color): Color {.inline.} =
+#   proc blend(backdrop, source: float32): float32 {.inline.} =
+#     backdrop + source - 2 * backdrop * source
+#   result.r = blend(backdrop.r, source.r)
+#   result.g = blend(backdrop.g, source.g)
+#   result.b = blend(backdrop.b, source.b)
+#   result = alphaFix(backdrop, source, result)
 
 proc blendColorFloats*(backdrop, source: Color): Color {.inline.} =
   result = SetLum(source, Lum(backdrop))
@@ -297,24 +297,24 @@ proc blendSaturationFloats*(backdrop, source: Color): Color {.inline.} =
   result = SetLum(SetSat(backdrop, Sat(source)), Lum(backdrop))
   result = alphaFix(backdrop, source, result)
 
-proc blendMaskFloats*(backdrop, source: Color): Color {.inline.} =
-  result = backdrop
-  result.a = min(backdrop.a, source.a)
+# proc blendMaskFloats*(backdrop, source: Color): Color {.inline.} =
+#   result = backdrop
+#   result.a = min(backdrop.a, source.a)
 
-proc blendSubtractMaskFloats*(backdrop, source: Color): Color {.inline.} =
-  result = backdrop
-  result.a = backdrop.a * (1 - source.a)
+# proc blendSubtractMaskFloats*(backdrop, source: Color): Color {.inline.} =
+#   result = backdrop
+#   result.a = backdrop.a * (1 - source.a)
 
-proc blendIntersectMaskFloats*(backdrop, source: Color): Color {.inline.} =
-  result = backdrop
-  result.a = backdrop.a * source.a
+# proc blendIntersectMaskFloats*(backdrop, source: Color): Color {.inline.} =
+#   result = backdrop
+#   result.a = backdrop.a * source.a
 
-proc blendExcludeMaskFloats*(backdrop, source: Color): Color {.inline.} =
-  result = backdrop
-  result.a = abs(backdrop.a - source.a)
+# proc blendExcludeMaskFloats*(backdrop, source: Color): Color {.inline.} =
+#   result = backdrop
+#   result.a = abs(backdrop.a - source.a)
 
-proc blendOverwriteFloats*(backdrop, source: Color): Color {.inline.} =
-  source
+# proc blendOverwriteFloats*(backdrop, source: Color): Color {.inline.} =
+#   source
 
 when defined(amd64) and not defined(pixieNoSimd):
   proc alphaFix(backdrop, source: ColorRGBA, vb, vs, vm: M128): ColorRGBA =