before:
```
name ............................... min time      avg time    std dv  times
blendNormal ........................ 0.882 ms      0.888 ms    ±0.018  x1000
blendDarken ........................ 1.074 ms      1.088 ms    ±0.037  x1000
blendMultiply ...................... 1.045 ms      1.054 ms    ±0.024  x1000
blendLinearBurn .................... 0.993 ms      1.015 ms    ±0.034  x1000
blendColorBurn ..................... 1.761 ms      1.778 ms    ±0.038  x1000
blendLighten ....................... 1.054 ms      1.076 ms    ±0.021  x1000
blendScreen ........................ 1.711 ms      1.725 ms    ±0.033  x1000
blendLinearDodge ................... 1.055 ms      1.062 ms    ±0.021  x1000
blendColorDodge .................... 1.725 ms      1.738 ms    ±0.032  x1000
blendOverlay ....................... 1.697 ms      1.717 ms    ±0.044  x1000
blendSoftLight ..................... 0.662 ms      0.667 ms    ±0.014  x1000
blendHardLight ..................... 1.694 ms      1.708 ms    ±0.034  x1000
blendDifference .................... 1.085 ms      1.094 ms    ±0.025  x1000
blendExclusion ..................... 1.710 ms      1.730 ms    ±0.045  x1000
blendHue ........................... 2.302 ms      2.326 ms    ±0.047  x1000
blendSaturation .................... 2.481 ms      2.508 ms    ±0.047  x1000
blendColor ......................... 1.843 ms      1.860 ms    ±0.037  x1000
blendLuminosity .................... 1.858 ms      1.876 ms    ±0.038  x1000
blendMask .......................... 0.094 ms      0.095 ms    ±0.003  x1000
blendSubtractMask .................. 0.118 ms      0.119 ms    ±0.004  x1000
blendIntersectMask ................. 0.092 ms      0.093 ms    ±0.004  x1000
blendExcludeMask ................... 0.102 ms      0.103 ms    ±0.004  x1000
blendNormalPremultiplied ........... 0.225 ms      0.226 ms    ±0.007  x1000
blendNormalPremultiplied [simd] .... 0.042 ms      0.043 ms    ±0.002  x1000
```

after:
```
name ............................... min time      avg time    std dv  times
blendNormal ........................ 0.607 ms      0.612 ms    ±0.015  x1000
blendDarken ........................ 0.472 ms      0.479 ms    ±0.015  x1000
blendMultiply ...................... 0.516 ms      0.530 ms    ±0.034  x1000
blendLinearBurn .................... 0.279 ms      0.281 ms    ±0.008  x1000
blendColorBurn ..................... 1.015 ms      1.027 ms    ±0.029  x1000
blendLighten ....................... 0.496 ms      0.501 ms    ±0.011  x1000
blendScreen ........................ 0.576 ms      0.580 ms    ±0.014  x1000
blendLinearDodge ................... 0.524 ms      0.529 ms    ±0.012  x1000
blendColorDodge .................... 0.428 ms      0.431 ms    ±0.011  x1000
blendOverlay ....................... 0.558 ms      0.572 ms    ±0.013  x1000
blendSoftLight ..................... 0.662 ms      0.667 ms    ±0.014  x1000
blendHardLight ..................... 0.333 ms      0.336 ms    ±0.009  x1000
blendDifference .................... 0.565 ms      0.570 ms    ±0.015  x1000
blendExclusion ..................... 0.568 ms      0.573 ms    ±0.013  x1000
blendHue ........................... 2.621 ms      2.662 ms    ±0.066  x1000
blendSaturation .................... 2.353 ms      2.391 ms    ±0.074  x1000
blendColor ......................... 1.920 ms      1.937 ms    ±0.036  x1000
blendLuminosity .................... 1.858 ms      1.878 ms    ±0.043  x1000
blendMask .......................... 0.094 ms      0.095 ms    ±0.004  x1000
blendSubtractMask .................. 0.106 ms      0.107 ms    ±0.004  x1000
blendIntersectMask ................. 0.092 ms      0.092 ms    ±0.003  x1000
blendExcludeMask ................... 0.102 ms      0.103 ms    ±0.005  x1000
blendNormalPremultiplied ........... 0.243 ms      0.245 ms    ±0.007  x1000
blendNormalPremultiplied [simd] .... 0.042 ms      0.043 ms    ±0.003  x1000
```